refactor: Move restaurant components to a dedicated folder and update imports
